Refactor mosrs.auth
