+++ mosrs/exception.py
@@ -1,0 +1,42 @@
+#!/usr/bin/env python
+"""
+Copyright 2015 ARC Centre of Excellence for Climate Systems Science
+
+author: Scott Wales <scott.wales@unimelb.edu.au>
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+class AuthError(Exception):
+    """
+    Indicates an anticipated error
+    """
+    pass
+
+class GPGError(Exception):
+    """
+    Indicates an anticipated error
+    """
+    pass
+
+class MessageError(Exception):
+    """
+    Indicates an anticipated error
+    """
+    pass
+
+class SetupError(Exception):
+    """
+    Indicates user needs to take action before setup can complete
+    """
+    pass
